+++ CEBNumericModel.h
